Use timerfd instead of a timeout for poll()
Screentime stats are now always rotated at midnight, even if the system entered suspend mode and exited: this is because a timerfd is now used instead of passing a timeout value to poll(). CLOCK_BOOTTIME is used when creating the timerfd, which makes it suspend-aware. Previously, stats rotation would be delayed by the suspend duration when the system suspended and resumed if the timer was never updated afterwards.

write_times() is now aware of whether stats should be archived or not, thus obsoleting save_times()